Move project files to WebQA directory
