+++ WebQA/Logic/Stat.cs
