+++ WebQA/Properties/AssemblyInfo.cs
